requirements.txt
telethon>=1.36
gate-api>=4.0
google-generativeai>=0.8
python-dotenv>=1.0
Pillow>=10.0
